Début du pattern Observer + Ajout de notre propre class Point.

+++ modele/src/main/java/Game/command/CommandMoveShape.java
@@ -5,9 +5,9 @@ import Game.model.Form.Shape;
 public class CommandMoveShape implements OperationCommand {
 
 	private Shape selectedShape;
-    private int dx, dy;
+    private double dx, dy;
 
-    public CommandMoveShape(Shape selectedShape, int dx, int dy) {
+    public CommandMoveShape(Shape selectedShape, double dx, double dy) {
         this.selectedShape = selectedShape;
         this.dx = dx;
         this.dy = dy;
