+++ modele/src/main/java/Game/model/Form/Shape.java
@@ -1,14 +1,14 @@
 package Game.model.Form;
 
-import java.awt.*;
-
+import java.awt.Rectangle;
+import Game.model.Point;
 import Game.model.collision.IntersectionVisiteur;
 
 public interface Shape {
     
 	Rectangle getBounds();
     
-    void move(int dx, int dy);
+    void move(double dx, double dy);
     
     boolean contains(Point p);
     
@@ -17,6 +17,5 @@ public interface Shape {
     void resize(Point lastPoint, Point newPoint); 
 
     void accept(IntersectionVisiteur  Visiteur);
-
     
 }
